Исправляет замечания после проверки наставником
