+++ js/form.js
@@ -1,6 +1,7 @@
 import { sendData } from './api.js';
 import { COORDS, addressInput, resetMap, closePopup } from './map.js';
 import { clearFilter } from './filter.js';
+import { clearFileUploadPreview } from './upload.js';
 
 const form = document.querySelector('.ad-form');
 const titleInput = form.querySelector('#title');
@@ -14,10 +15,7 @@ const submitButton = form.querySelector('.ad-form__submit');
 const resetButton = form.querySelector('.ad-form__reset');
 const successMessageTemplate = document.querySelector('#success').content.querySelector('.success');
 const errorMessageTemplate = document.querySelector('#error').content.querySelector('.error');
-const avatarChooser = form.querySelector('#avatar');
-const avatarPreview = form.querySelector('.ad-form-header__preview');
-const imagesChooser = form.querySelector('#images');
-const imagesPreview = form.querySelector('.ad-form__photo');
+
 let message;
 
 // Функция для проверки количества комнат и гостей
@@ -64,6 +62,7 @@ function enableForm () {
 function clearForm () {
   form.reset();
   addressInput.value = `${COORDS.lat}, ${COORDS.lng}`;
+  clearFileUploadPreview();
 }
 
 // Получаем минимальное значение в поле "Цена за ночь"
@@ -190,7 +189,10 @@ function setFormSubmit () {
 
     const formData = new FormData(evt.target);
 
-    sendData(showSuccessMessage, showErrorMessage, formData);
+    sendData(() => {
+      showSuccessMessage();
+      clearForm();
+    }, showErrorMessage, formData);
   });
 }
 
@@ -240,37 +242,4 @@ function showErrorMessage () {
   document.body.appendChild(message);
 }
 
-const FILE_TYPES = ['gif', 'jpg', 'jpeg', 'png'];
-
-function setFileUploadPreview (chooser, preview) {
-  let previewImage = preview.querySelector('img');
-
-  chooser.addEventListener('change', () => {
-    const file = chooser.files[0];
-    const fileName = file.name.toLowerCase();
-
-    const matches = FILE_TYPES.some((item) => fileName.endsWith(item));
-
-    if (matches) {
-      if (previewImage) {
-        previewImage.style.width = '70px';
-        previewImage.style.height = '70px';
-        previewImage.style.borderRadius = '5px';
-        previewImage.src = URL.createObjectURL(file);
-      } else {
-        previewImage = document.createElement('img');
-        previewImage.style.width = '70px';
-        previewImage.style.height = '70px';
-        previewImage.style.borderRadius = '5px';
-        previewImage.src =  URL.createObjectURL(file);
-        preview.append(previewImage);
-      }
-    }
-  });
-}
-
-setFileUploadPreview(avatarChooser, avatarPreview);
-setFileUploadPreview (imagesChooser, imagesPreview);
-
-
-export { disableForm, enableForm, clearForm, setFormSubmit };
+export { disableForm, enableForm, clearForm, setFormSubmit, form };
